Rename to announce.rs and delete.rs
