+++ contracts/service-provider-directory/src/integration_tests/delete.rs
